ref: Restructure Rust code into clear modules
