+++ tests/resolver.rs
